requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
bcrypt==4.0.1
black==23.7.0
blinker==1.6.2
CacheControl==0.13.1
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.6
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
Cython==0.29.33
dataclasses==0.6
fastapi==0.87.0
filetype==1.2.0
fire==0.5.0
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-Session==0.5.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
ImageHash==4.3.1
imageio==2.25.1
imantics==0.1.12
imgaug==0.4.0
imutils==0.5.4
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.13.1
kiwisolver==1.4.4
labelme2coco==0.2.3
libclang==15.0.6.1
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.13.1
opencv-python==4.7.0.72
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.0
Pillow==9.3.0
pip==23.2.1
pixellib==0.7.1
platformdirs==3.0.0
pooch==1.6.0
portalocker==2.7.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybboxes==0.1.6
pycparser==2.21
pydantic==1.10.5
pydot==1.4.2
PyJWT==2.8.0
PyMatting==1.1.8
PyMySQL==1.0.2
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
rembg==2.0.50
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sahi==0.11.14
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.9.3
setuptools==65.5.1
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.21.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.2.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
watchdog==2.1.9
Werkzeug==2.2.3
wrapt==1.14.1
xmljson==0.2.1
yacs==0.1.8
yarl==1.8.2
